+++ elisp/src/main/java/party/iroiro/juicemacs/elisp/forms/regex/ELispRegExpInputNodes.java
@@ -16,30 +16,30 @@ abstract class ELispRegExpInputNodes {
 
     @GenerateInline(value = false)
     abstract static class InputLengthNode extends Node {
-        public abstract int execute(VirtualFrame frame, Object input);
+        public abstract long execute(VirtualFrame frame, Object input);
 
         @Specialization
-        public int testInputStrLength(MuleString input) {
+        public long testInputStrLength(MuleString input) {
             return input.length();
         }
 
         @Specialization
-        public int testInputBufferLength(ELispBuffer input) {
+        public long testInputBufferLength(ELispBuffer input) {
             return input.length();
         }
     }
 
     @GenerateInline(value = false)
     abstract static class InputGetCharNode extends Node {
-        public abstract int execute(VirtualFrame frame, Object input, int index);
+        public abstract int execute(VirtualFrame frame, Object input, long index);
 
         @Specialization
-        public int inputGetStrChar(MuleString input, int index) {
+        public int inputGetStrChar(MuleString input, long index) {
             return input.codePointAt(index);
         }
 
         @Specialization
-        public int inputGetBufferChar(ELispBuffer input, int index) {
+        public int inputGetBufferChar(ELispBuffer input, long index) {
             return (int) input.getChar(index);
         }
     }
